requirements.txt
Flask>=3.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0

